+++ CHANGELOG.md
@@ -18,6 +18,8 @@ The minor version will be incremented upon a breaking change and the patch versi
 
 ### Breaking
 
+- tools: add metrics, new config for google-pubsub ([#279](https://github.com/rpcpool/yellowstone-grpc/pull/279))
+
 ## 2024-01-15
 
 - yellowstone-grpc-client-1.13.0+solana.1.17.16